Move stuff and make new file for funcs that take in and return FOOOF objects
