+++ fooof/__init__.py
@@ -3,4 +3,5 @@
 from .version import __version__
 
 from .fit import FOOOF
-from .group import FOOOFGroup, fit_fooof_group_3d
+from .group import FOOOFGroup
+from .funcs import fit_fooof_group_3d
